Fixed the documentation for parameter_key
The old structure wasn't allowing the documentation of subclasses to
override their parent classes, so the property has been restructured.

The recommended way of combining the abc library with the @property
decorator has a bug that doesn't allow the child class to override the
parent properties, so this property is simply documented that it should
be overridden.