Add solution bounds to the box constraints (#185)
The upper and lower limit of the box constraints were being implicitly
factored into the equations as though the lower limit was 0 and the
upper limit was 1. To make the code more versatile, we've made the
equations explicit and added the solution_bounds as a parameter to the
ProblemInstance class.